Fix based on comments
