+++ core/src/main/java/com/scalar/db/transaction/consensuscommit/ConsensusCommitUtils.java
@@ -5,6 +5,7 @@ import com.scalar.db.api.TableMetadata;
 import com.scalar.db.io.DataType;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -28,7 +29,7 @@ public final class ConsensusCommitUtils {
           .put(Attribute.COMMITTED_AT, DataType.BIGINT)
           .build();
 
-  static final ImmutableMap<String, DataType> TRANSACTION_META_COLUMNS =
+  private static final ImmutableMap<String, DataType> TRANSACTION_META_COLUMNS =
       ImmutableMap.<String, DataType>builder()
           .putAll(AFTER_IMAGE_META_COLUMNS)
           .putAll(BEFORE_IMAGE_META_COLUMNS)
@@ -127,11 +128,25 @@ public final class ConsensusCommitUtils {
     return true;
   }
 
-  static Set<String> getNonPrimaryKeyColumns(TableMetadata tableMetadata) {
+  /**
+   * Get non-primary key columns from the specified table metadata.
+   *
+   * @return a set of non-primary key column names
+   */
+  public static Set<String> getNonPrimaryKeyColumns(TableMetadata tableMetadata) {
     return tableMetadata.getColumnNames().stream()
         .filter(c -> !tableMetadata.getPartitionKeyNames().contains(c))
         .filter(c -> !tableMetadata.getClusteringKeyNames().contains(c))
         .collect(Collectors.toSet());
+  }
+
+  /**
+   * Get transaction meta columns.
+   *
+   * @return a map of transaction meta columns
+   */
+  public static Map<String, DataType> getTransactionMetaColumns() {
+    return TRANSACTION_META_COLUMNS;
   }
 
   /**
