+++ core/src/main/java/com/scalar/db/transaction/jdbc/JdbcTransactionAdmin.java
@@ -1,7 +1,6 @@
 package com.scalar.db.transaction.jdbc;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.ImmutableMap;
 import com.google.inject.Inject;
 import com.scalar.db.api.DistributedStorageAdmin;
 import com.scalar.db.api.DistributedTransactionAdmin;
@@ -93,10 +92,7 @@ public class JdbcTransactionAdmin implements DistributedTransactionAdmin {
 
   @Override
   public void importTable(String namespace, String table) throws ExecutionException {
-    TableMetadata tableMetadata = jdbcAdmin.getImportTableMetadata(namespace, table);
-
-    // add ScalarDB metadata
-    jdbcAdmin.repairTable(namespace, table, tableMetadata, ImmutableMap.of());
+    jdbcAdmin.importTable(namespace, table);
   }
 
   @Override
